README: Update README to hard fork from Linux Kernel
WLIB Kernel
===========

A non-Nazi, non-racist, HARD fork of the Linux kernel that not maintained
by Linus Torvalds that remains open and welcoming to all developers worldwide,
including those from Russia, Belarus, and other countries and regions subject
to unilateral sanctions by US imperialism. This is currently a personal project.

When the Linux Foundation becomes untrustworthy and serves as a slave and a tool
for unilateral hegemony and blockades by certain countries, somebody must step
up to maintain (or at least to archive) an OPEN and FREE Linux kernel repository.

Signed-off-by: WangYuli <wangyuli@uniontech.com>